feat: add a class to detect if ArcDPS has timed out or not
This commit adds a class detecting if ArcDPS has timed out or not.
It measures the interval of time since the last ArcDPS event, and declares a
time out after 60 seconds.

After the time out, the counter is reset, and a message is logged.

To prevent it from triggering too often when idle, it only counts when the
player is in combat. The only way I've managed to trigger a false positive
is to have a totally empty build on an armorless (because of 6th rune bonus)
character, and being idle after hitting a training golem.

I think it should be okay, we'll see with usage!